tests/wpt/tests/content-security-policy/worker-src/shared-none.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="worker-src 'none'">
<script>
  var url = new URL("../support/ping.js", document.baseURI).toString();
  assert_shared_worker_is_blocked(url, "Same-origin shared worker blocked by 'none'.");

  var b = new Blob(["onconnect = e => { e.ports[0].postMessage('ping'); }"], {type: "text/javascript"});
  var url = URL.createObjectURL(b);
  assert_shared_worker_is_blocked(url, "blob: shared worker blocked by 'none'.");
</script>
