LayoutTests/imported/w3c/web-platform-tests/content-security-policy/worker-src/shared-none.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.


FAIL Same-origin shared worker blocked by 'none'. assert_throws_dom: function "function () {
      var w = new SharedWorker(url);
    }" threw object "ReferenceError: Can't find variable: SharedWorker" that is not a DOMException SecurityError: property "code" is equal to undefined, expected 18
FAIL blob: shared worker blocked by 'none'. assert_throws_dom: function "function () {
      var w = new SharedWorker(url);
    }" threw object "ReferenceError: Can't find variable: SharedWorker" that is not a DOMException SecurityError: property "code" is equal to undefined, expected 18
